+++ core/client/controllers/editor.js
@@ -1,0 +1,5 @@
+import EditorControllerMixin from 'ghost/mixins/editor-base-controller';
+
+var EditorController = Ember.ObjectController.extend(EditorControllerMixin);
+
+export default EditorController;
