+++ core/client/routes/editor.js
@@ -3,7 +3,7 @@ import AuthenticatedRoute from 'ghost/routes/authenticated';
 
 var EditorRoute = AuthenticatedRoute.extend(styleBody, {
     classNames: ['editor'],
-    controllerName: 'posts.post',
+
     model: function (params) {
         var post = this.store.getById('post', params.post_id);
 
@@ -11,7 +11,7 @@ var EditorRoute = AuthenticatedRoute.extend(styleBody, {
             return post;
         }
 
-        return this.store.filter('post', { status: 'all' }, function (post) {
+        return this.store.filter('post', { status: 'all', staticPages: 'all' }, function (post) {
             return post.get('id') === params.post_id;
         }).then(function (records) {
             return records.get('firstObject');
