+++ core/client/templates/-floating-header.hbs
@@ -18,7 +18,7 @@
             <span class="hidden">Post Settings</span>
         {{/gh-popover-button}}
         {{#gh-popover name="post-settings-menu" classNames="post-settings-menu menu-drop-right"}}
-            {{view "post-settings-menu-view"}}
+            {{render "post-settings-menu" model}}
         {{/gh-popover}}
     </section>
 </header>
