+++ core/client/templates/-publish-bar.hbs
@@ -9,7 +9,7 @@
                     <span class="hidden">Post Settings</span>
                 {{/gh-popover-button}}
                 {{#gh-popover name="post-settings-menu" classNames="post-settings-menu menu-right"}}
-                    {{view "post-settings-menu-view"}}
+                    {{render "post-settings-menu" model}}
                 {{/gh-popover}}
             </section>
