--- core/client/templates/new.hbs
@@ -1,1 +1,0 @@
-TODO
