+++ core/client/views/new.js
@@ -1,7 +1,8 @@
-var EditorView = Ember.View.extend({
+var NewView = Ember.View.extend({
     tagName: 'section',
+    templateName: 'editor',
     classNames: ['entry-container'],
     scrollPosition: 0  // percentage of scroll position
 });
 
-export default EditorView;
+export default NewView;
